update and break up services/yaml tests
