+++ src/services/yaml/useYamlConfig.spec.jsx
@@ -1,10 +1,10 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { renderHook, waitFor } from '@testing-library/react'
-import { rest } from 'msw'
-import { setupServer } from 'msw/node'
+import { http, HttpResponse } from 'msw2'
+import { setupServer } from 'msw2/node'
 import { MemoryRouter, Route } from 'react-router-dom'
 
-import { useUpdateYaml, useYamlConfig } from './index'
+import { useYamlConfig } from './useYamlConfig'
 
 const queryClient = new QueryClient({
   defaultOptions: { queries: { retry: false } },
@@ -19,18 +19,24 @@ const wrapper = ({ children }) => (
 
 const server = setupServer()
 
-beforeAll(() => server.listen())
+beforeAll(() => {
+  server.listen()
+})
+
 beforeEach(() => {
   server.resetHandlers()
   queryClient.clear()
 })
-afterAll(() => server.close())
+
+afterAll(() => {
+  server.close()
+})
 
 describe('useYamlConfig', () => {
   function setup(dataReturned = {}) {
     server.use(
-      rest.post(`/graphql/gh`, (req, res, ctx) => {
-        return res(ctx.status(200), ctx.json({ data: dataReturned }))
+      http.post('/graphql/gh', (info) => {
+        return HttpResponse.json({ data: dataReturned })
       })
     )
   }
@@ -74,57 +80,6 @@ describe('useYamlConfig', () => {
       await waitFor(() => !result.current.isLoading)
 
       await waitFor(() => expect(result.current.data).toEqual('hello'))
-    })
-  })
-})
-
-describe('useUpdateYaml', () => {
-  function setup(dataReturned = {}) {
-    server.use(
-      rest.post(`/graphql/gh`, (req, res, ctx) => {
-        return res(ctx.status(200), ctx.json(dataReturned))
-      })
-    )
-  }
-
-  describe('when mutate is called', () => {
-    describe('and is authenticated', () => {
-      beforeEach(() => {
-        jest.resetAllMocks()
-        setup({
-          data: {
-            setYamlOnOwner: {
-              owner: { yaml: 'hello: there', username: 'doggo' },
-            },
-          },
-        })
-      })
-
-      it('to return the new yaml', async () => {
-        const { result } = renderHook(
-          () => useUpdateYaml({ username: 'doggo' }),
-          {
-            wrapper,
-          }
-        )
-
-        result.current.mutate({
-          yaml: 'hello:there',
-        })
-
-        await waitFor(() => result.current.isLoading)
-        await waitFor(() => !result.current.isLoading)
-
-        await waitFor(() =>
-          expect(result.current.data).toStrictEqual({
-            data: {
-              setYamlOnOwner: {
-                owner: { yaml: 'hello: there', username: 'doggo' },
-              },
-            },
-          })
-        )
-      })
     })
   })
 })
